move docs to basic
